Replace logic for patch files with proper branch merging.
The "git am" command was rather fragile, failing when trying to apply
to a newer version (v1.19.1) that already had the most up-to-date JS
file.  So drop the patch files here and maintain the needed commits
for "backporting" in the docs repo itself, as a separate branch.
Using "git merge" on that allows better control of the merge strategy
and succeeds in more possible scenarios.